Move shamir's into public sdk/ space, fix titles
This moves shamir's code into the public sdk/ space, encouraging the
reuse of this component by other external libraries. Also fixes the
titles of API and SDK READMEs.
