+++ api/README.md
@@ -1,9 +1,10 @@
-Vault API
+OpenBao API
 =================
 
-This provides the `github.com/openbao/openbao/api` package which contains code useful for interacting with a Vault server.
+This provides the `github.com/openbao/openbao/api` package which contains code useful for interacting with an OpenBao or Vault server.
 
-For examples of how to use this module, see the [vault-examples](https://github.com/hashicorp/vault-examples) repo.
-For a step-by-step walkthrough on using these client libraries, see the [developer quickstart](https://www.vaultproject.io/docs/get-started/developer-qs).
+For examples of how to use this module, see the test suite contained
+within OpenBao or [vault-examples](https://github.com/hashicorp/vault-examples)
+repo from HashiCorp.
 
 [![GoDoc](https://godoc.org/github.com/openbao/openbao/api?status.png)](https://godoc.org/github.com/openbao/openbao/api)
