+++ changelog/181.txt
@@ -1,0 +1,3 @@
+```release-note:improvement
+sdk/helper/shamir: move Shamir's code into public SDK namespace to encourage external reuse
+```
