+++ sdk/README.md
@@ -1,12 +1,12 @@
-Vault SDK libs
+OpenBao SDK libs
 =================
 
 This package provides the `sdk` package which contains code useful for
-developing Vault plugins.
+developing OpenBao plugins.
 
 Although we try not to break functionality, we reserve the right to reorganize
 the code at will and may occasionally cause breaks if they are warranted. As
 such we expect the tag of this module will stay less than `v1.0.0`.
 
 For any major changes we will try to give advance notice in the CHANGES section
-of Vault's CHANGELOG.md.
+of OpenBao's CHANGELOG.md.
