+++ vault/core.go
@@ -53,10 +53,10 @@ import (
 	"github.com/openbao/openbao/sdk/helper/jsonutil"
 	"github.com/openbao/openbao/sdk/helper/logging"
 	"github.com/openbao/openbao/sdk/helper/pathmanager"
+	"github.com/openbao/openbao/sdk/helper/shamir"
 	"github.com/openbao/openbao/sdk/logical"
 	"github.com/openbao/openbao/sdk/physical"
 	sr "github.com/openbao/openbao/serviceregistration"
-	"github.com/openbao/openbao/shamir"
 	"github.com/openbao/openbao/vault/cluster"
 	"github.com/openbao/openbao/vault/eventbus"
 	"github.com/openbao/openbao/vault/quotas"
